+++ src/main/java/Model/Entitys/Monsters/Alien.java
@@ -6,7 +6,7 @@ import Model.Utils.Position;
 import Model.Utils.Procedure;
 
 public class Alien extends AbstractMonster{
-    public Alien(Etage m, Position pos, String nom, int pv, int force, int vision_radius , int agro, int update_rate, int pathCross) {
+    public Alien(Etage m, Position pos, String nom, int pv, int force, double vision_radius , int agro, int update_rate, int pathCross) {
         super(m, pos, nom, pv, force, vision_radius, agro, update_rate, pathCross);
     }
