+++ src/main/java/Model/Entitys/Monsters/Bee.java
@@ -7,7 +7,7 @@ import Model.Utils.Position;
 public class Bee extends AbstractMonster{
     private final int BasePv;
 
-    public Bee(Etage m, Position pos, String nom, int pv, int force, int vision_radius , int agro, int update_rate, int pathCross) {
+    public Bee(Etage m, Position pos, String nom, int pv, int force, double vision_radius , int agro, int update_rate, int pathCross) {
         super(m, pos, nom, pv, force, vision_radius, agro, update_rate, pathCross);
         this.BasePv=pv;
     }
