+++ src/main/java/Model/Entitys/Monsters/Snail.java
@@ -6,7 +6,7 @@ import Model.Utils.Position;
 
 public class Snail extends AbstractMonster {
     private boolean toleft = true;
-    public Snail(Etage m, Position pos, String nom, int pv, int force, int vision_radius , int agro, int update_rate, int pathCross) {
+    public Snail(Etage m, Position pos, String nom, int pv, int force, double vision_radius , int agro, int update_rate, int pathCross) {
         super(m, pos, nom, pv, force, vision_radius, agro, update_rate, pathCross);
     }
